devc-model-master/requirements.txt
Django==3.1.3
django-cors-headers==3.5.0
django-heroku==0.3.1
djangorestframework==3.12.2
djangorestframework-jwt==1.11.0
djongo==1.3.3
dnspython==2.0.0
facebook-sdk==3.1.0
gensim==3.8.3
gunicorn==20.0.4
importlib-metadata==1.5.0
jsonschema==3.2.0
nltk==3.5
numpy==1.18.5
pandas==1.1.2
pickle-mixin==1.0.2
pickleshare==0.7.5
PyJWT==1.7.1
pymongo==3.11.0
python-dateutil==2.7.3
python-facebook-api==0.7.1
pytz==2019.3
redis==3.5.3
requests==2.24.0
requests-oauthlib==1.3.0
requests-unixsocket==0.2.0
rq==1.6.1
sqlparse==0.2.4
https://download.pytorch.org/whl/cpu/torch-1.6.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
whitenoise==5.2.0
xlrd==1.2.0
